+++ tests/curd.test.ts
@@ -1,6 +1,7 @@
-import { MongoClient } from "./src/client.ts";
-import { ObjectId } from "./src/utils/bson.ts";
+import { MongoClient } from "../src/client.ts";
 import { assert, assertEquals, assertThrowsAsync } from "./test.deps.ts";
+import { testWithClient } from "./common.ts";
+
 interface IUser {
   username: string;
   password: string;
@@ -10,23 +11,6 @@ interface IUser {
 const { test } = Deno;
 const dateNow = Date.now();
 const hostName = "localhost";
-
-async function testWithClient(
-  name: string,
-  fn: (client: MongoClient) => void | Promise<void>,
-) {
-  test(name, async () => {
-    const client = await getClient();
-    await fn(client);
-    client.close();
-  });
-}
-
-async function getClient(): Promise<MongoClient> {
-  const client = new MongoClient();
-  await client.connect(`mongodb://${hostName}:27017`);
-  return client;
-}
 
 test("testConnectWithUri", async () => {
   const client = new MongoClient();
@@ -87,7 +71,7 @@ testWithClient("testUpsertOne", async (client) => {
       password: "pass1",
       date: new Date(dateNow),
     },
-    { upsert: true },
+    { upsert: true }
   );
 
   assert(upsertedId);
@@ -119,7 +103,7 @@ testWithClient("testInsertOneTwice", async (client) => {
         username: "user1",
       }) as any,
     undefined,
-    "E11000",
+    "E11000"
   );
 });
 
@@ -134,12 +118,12 @@ await testWithClient("testFindOne", async (client) => {
   assertEquals(findNull, undefined);
   const projectionUser = await users.findOne(
     {},
-    { projection: { _id: 0, username: 1 } },
+    { projection: { _id: 0, username: 1 } }
   );
   assertEquals(Object.keys(projectionUser!), ["username"]);
   const projectionUserWithId = await users.findOne(
     {},
-    { projection: { username: 1 } },
+    { projection: { username: 1 } }
   );
   assertEquals(Object.keys(projectionUserWithId!), ["_id", "username"]);
 });
@@ -187,7 +171,7 @@ testWithClient("testUpdateOneWithUpsert", async (client) => {
   const result = await users.updateOne(
     { username: "user2" },
     { username: "USER2" },
-    { upsert: true },
+    { upsert: true }
   );
   assertEquals(result.matchedCount, 1);
   assertEquals(result.modifiedCount, 0);
@@ -250,7 +234,7 @@ testWithClient("testUpdateMany", async (client) => {
   const users = db.collection("mongo_test_users");
   const result = await users.updateMany(
     { username: "many" },
-    { $set: { username: "MANY" } },
+    { $set: { username: "MANY" } }
   );
   assertEquals(result, {
     matchedCount: 1,
