+++ tests/utils/uri.test.ts
@@ -1,5 +1,5 @@
 import { parse } from "../../src/utils/uri.ts";
-import { assertEquals } from "../../test.deps.ts";
+import { assertEquals } from "../test.deps.ts";
 
 Deno.test({
   name: "should correctly parse mongodb://localhost",
@@ -28,7 +28,7 @@ Deno.test({
     "should correctly parse mongodb://localhost:27017test?appname=hello%20world",
   fn() {
     const options = parse(
-      "mongodb://localhost:27017/test?appname=hello%20world",
+      "mongodb://localhost:27017/test?appname=hello%20world"
     );
     assertEquals(options.appname, "hello world");
   },
@@ -39,7 +39,7 @@ Deno.test({
     "should correctly parse mongodb://localhost/?safe=true&readPreference=secondary",
   fn() {
     const options = parse(
-      "mongodb://localhost/?safe=true&readPreference=secondary",
+      "mongodb://localhost/?safe=true&readPreference=secondary"
     );
     assertEquals(options.dbName, "admin");
     assertEquals(options.servers.length, 1);
@@ -51,9 +51,7 @@ Deno.test({
 Deno.test({
   name: "should correctly parse mongodb://localhost:28101/",
   fn() {
-    const options = parse(
-      "mongodb://localhost:28101/",
-    );
+    const options = parse("mongodb://localhost:28101/");
     assertEquals(options.dbName, "admin");
     assertEquals(options.servers.length, 1);
     assertEquals(options.servers[0].host, "localhost");
